+++ src/components/decoration4/src/main.vue
@@ -2,7 +2,7 @@
   <div class="dv-decoration-4" :ref="ref">
     <div
       :class="`container ${reverse ? 'reverse' : 'normal'}`"
-      :style="reverse ? `width:${width}px;height:5px` : `width:5px;height:${height}px;`"
+      :style="reverse ? `width:${width}px;height:5px;animation-duration:${dur}s` : `width:5px;height:${height}px;animation-duration:${dur}s`"
     >
       <svg :width="reverse ? width : 5" :height="reverse ? 5 : height">
         <polyline
@@ -40,6 +40,10 @@ export default {
     reverse: {
       type: Boolean,
       default: false
+    },
+    dur: {
+      type: Number,
+      default: 3
     }
   },
   data () {
@@ -83,23 +87,26 @@ export default {
     display: flex;
     overflow: hidden;
     position: absolute;
+    flex: 1;
   }
 
   .normal {
-    height: 0% !important;
-    animation: ani-height 3s ease-in-out infinite;
+    animation: ani-height ease-in-out infinite;
     left: 50%;
     margin-left: -2px;
   }
 
   .reverse {
-    width: 0% !important;
-    animation: ani-width 3s ease-in-out infinite;
+    animation: ani-width ease-in-out infinite;
     top: 50%;
     margin-top: -2px;
   }
 
   @keyframes ani-height {
+    0% {
+      height: 0%;
+    }
+
     70% {
       height: 100%;
     }
@@ -110,6 +117,10 @@ export default {
   }
 
   @keyframes ani-width {
+    0% {
+      width: 0%;
+    }
+    
     70% {
       width: 100%;
     }